Abstracting chart into more general component
