+++ src/explore-education-statistics-common/src/modules/find-statistics/components/charts/Charts.ts
@@ -1,6 +1,3 @@
-import { Axis, ReferenceLine } from '@common/services/publicationService';
-import { CharacteristicsData } from '@common/services/tableBuilderService';
-
 export const colours: string[] = [
   '#b10e1e',
   '#006435',
@@ -21,14 +18,4 @@ export function parseCondensedTimePeriodRange(
   separator: string = '/',
 ) {
   return [range.substring(0, 4), range.substring(4, 6)].join(separator);
-}
-
-export interface ChartProps {
-  characteristicsData: CharacteristicsData;
-  chartDataKeys: string[];
-  labels: { [key: string]: string };
-  xAxis: Axis;
-  yAxis: Axis;
-  height?: number;
-  referenceLines?: ReferenceLine[];
 }
