+++ src/explore-education-statistics-common/src/modules/find-statistics/components/charts/MapBlock.tsx
@@ -3,8 +3,7 @@ import { LatLngBounds } from 'leaflet';
 import 'leaflet/dist/leaflet.css';
 import React, { Component } from 'react';
 import { GeoJSON, Map } from 'react-leaflet';
-
-import { ChartProps } from './Charts';
+import { ChartProps } from '@common/modules/find-statistics/components/charts/AbstractChart';
 
 export type MapFeature = Feature<Geometry, GeoJsonProperties>;
